Fix warnings on windows

+++ cpp-terminal/buffer.cpp
@@ -48,7 +48,7 @@ void Term::Buffer::setType(const Term::Buffer::Type& type) { m_type = type; }
 
 std::streambuf* Term::Buffer::setbuf(char* s, std::streamsize n)
 {
-  if(s != nullptr) m_buffer.reserve(n);
+  if(s != nullptr) m_buffer.reserve(static_cast<std::size_t>(n));
   return this;
 }
